fix #700
Added new argument to `check_to_html` that will display a certain text when the check status is `FALSE`. This `FALSE` status for a check is setted when a cross-check (e.g. samples-counts) has not been performed due to one of the tables not being uploaded.

Also allow sequential data upload by not overwritting to NULL all the tables when a new upload is performed.

For sequential also do not overload the checklist with the checklists of ONLY the latest upload, we want to keep the previous checklist results. If we upload sequentially 2 counts.csv the results will be overwritten with the latest counts update though, so we are safe.

We also now allow to upload just the contrasts table without the counts or samples present, the user will need to upload them to move forward so there is no risk.