fix: sidebar issues
there were a couple of issues on the sidebar loading/unloading
procedure. First some elements were not removed properly when refreshind
the app (first dropdown elements were vissible); then, the first
dropdown was sometimes stuck (with no elements displayed) and appeared
with the down arrow as it was opened. Both issues are fixed, maybe it is
a little bit ugly though