feat: add req for data change
when changing datasets, at a point on the reactivity triggered by it,
the selected gene is still from the old data, which can be that is not
found on the new dataset. that caused an error; by adding this req we
avoid it. 1) no split second error message is displayed on the platform
2) no ticket is created